debug
Typing large number in input, the error occurs, so first I try to add
maxlength attribute into input tag, but that tag is for ‘type=“text’,
so I switch to min and max attribute, however, I found it won’t
validate until sending the form. So I reuse the maxlength way, and use
isNaN as the validation.